Finalised work on the combat tracker
Using skills in and out of combat or the actors turn apply the heart
rate change in a correct way and notify via the chat if necessary.
The heart rate changes are applied correctly, taking thresholds into
account and resting rounds. At the end of a turn the heart rate changes
are automatically applied.